add class in pipeline file
