--- src/org/maneyko/MyClass.groovy
@@ -1,9 +1,0 @@
-package org.maneyko
-
-class MyClass {
-    String  var1  = "var1"
-    String  var2  = "var2"
-    String  var3  = "var3"
-    String  var4  = "var4"
-    Boolean bool1 = true
-}
